가계부 > 카드 사용 내역 검색 기능 수정

+++ code/src/main/java/com/project/accountbook/account/model/AccountInfoDTO.java
@@ -23,7 +23,7 @@ public class AccountInfoDTO {
     /**
      * 가격입니다.
      */
-    private int price;
+    private long price;
     
     /**
      * 위치입니다.
@@ -73,12 +73,12 @@ public class AccountInfoDTO {
     /**
      * 총 가격입니다.
      */
-    private int totalPrice;
+    private long totalPrice;
     
     /**
      * 총 저축액입니다.
      */
-    private int totalSaving;
+    private long totalSaving;
 
     //가계부
     /**
